+++ man-roxygen/param.autopool.R
@@ -1,0 +1,2 @@
+#' @param autopool Should the automatic pooling algorithms be used.
+#' Give more details here on these rule work.